// src/shared/types/accountwiseSummaryReport.ts
export type AccountwiseSummaryRecord = {
  id: string
  accountName: string
  otherName: string
  mobileNumber: string
  city: string
  groupName: string
  totalSaleFine: number
  totalSaleValue: number
  totalPurchaseFine: number
  totalPurchaseValue: number
  goldFine: number
  silverFine: number
  cash: number
  anamat: number
  bank: number
}